// src/api.ts
export interface TreeNode {
  id: string;
  type: "project" | "folder" | "file";
  name: string;
  children?: TreeNode[];
}